+++ package.json
@@ -29,6 +29,7 @@
     "next-auth": "^5.0.0-beta.20",
     "react": "^18",
     "react-dom": "^18",
+    "react-hot-toast": "^2.4.1",
     "react-loading-skeleton": "^3.4.0",
     "swr": "^2.2.5",
     "tailwind-merge": "^2.5.2",
