+++ src/app/layout.tsx
@@ -29,8 +29,8 @@ const fontBody = Space_Mono({
 
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "MindR",
+  description: "Never loose any memory again",
 };
 
 export default function RootLayout({
