+++ src/components/LoginButton.tsx
@@ -45,7 +45,7 @@ const TelegramLoginButton: React.FC = () => {
     const onAuth = async (user: any) => {
             try {
               console.log('Telegram user data:', user);
-              const object = await signIn('credentials', { ...user, callback: '/dashboard', redirect:false });
+              const object = await signIn('credentials', { ...user, callback: '/dashboard', redirect:true });
               console.log('SignIn result:', object);
               // Handle successful sign-in here 
             } catch (error) {
